Version 1.9.1 Under Beta!
Version 1.9.0 Release:
-2 New Commands were added!
-b/report - Reports a user in that discord server. You need to have a #reports channel, or else, the bot will not send it through.
-b/suggest - Suggest something for that discord server. You need to have a #suggestions channel, or else, the bot will not send it through.

b/help section
-All of the sections are now spaced out, so it doesn't feel congested.
-b/report was added to Moderation Commands in the help section
-b/suggest was added to the Interactive Commands in the help section

Coding
The new file was installed called .env. There are many problems with it right now, and planning to do something with it later. Right now, the bot is still functioning, but there is a few security risks right now we are adding. I cannot see who is using what command, or who is using it in their discord server (unless if I am in it). That is all. It is not shown here yet, because it is still under works, and there are a few bugs, I am trying to fix.

Version 1.9.1 Release:
-In the command b/permissions, it said Moderaiton, instead of moderation. My bad. Dumb mistake.

More updates coming soon...