package.json info
This is where all the presets where my files are stored. This was created when I made the npm file in command prompt. All of this was set up manually when making the .json file.
I added some // text and some info about this file, to make it a bit easier to read for new comers.