name.json info
All of the .json stuff here will come from the test.js file, where the code will run. It looks like this now, because I have been working a bit with XShadowWizardX (Ranked #3 with the most stars in Geometry Dash) and we finished up the first steps. The Values are not defined yet and will be in the future...