Version 1.91.3 Under Beta! (Index.js code)
WE TOTALLY FORGOT TO MENTION THE NEW CODE WE ARE WRITING FOR b/suggest MY BAD -___- SORRY.
You can find this at the bottom of the code. It is all in messages. I explained a bit in detail on how that is support to work. Well, kind of...

Version 1.91.2 Release:
-Made a small adjustment to b/info about. discord.js version now says <BOLDEED(discord.js version)>. That is the only change we made. 

Version 1.91.3 Release:
-In the b/help command, there wasn't a dash where the b/say was. It is now 100% fixed. Sorry -__- (was suppose to be -b/say in b/help. It said b/say in b/help)

More updates are coming soon...