Version 1.0.0 Stage Alpha
Yes, we have made it to Stage Alpha! After over a month, the bot is finally in Stage Alpha!!!
Here is what is new in this release

-The bot now only responds to b/ prefix instead of 2 characters.
We added this in the code:
if(!message.content.startsWith(PREFIX) || message.author.bot)return; //This is what filters out the prefix. Without this here, the bot would still be able to respond to 2 characters
    let args = message.content.substring(PREFIX.length).split (/ +/); //Split was changed to this. Doesn't really affect anything to be honest, but it works.

-b/info about and b/info versions were updated. The variable for both of those commands are updated too.

More updates are coming soon...