Version 1.02.0 Stage Alpha
I did not have my PC to update the bot code on index.js. Here are all the releases starting off with 1.01.0.

*Version 1.01.0 Release*
-Updated b/clear. When doing command, it felt like a little bit rude, and a bit angry. We tried to make it nicer, and a bit more informal.
-b/ban and b/kick have an update!
For this update, we fixed 1 message that was annoying. When doing b/kick and b/ban in general, it will still show the message, but if you do b/kick <user>, that message will not show again. 
(b/kick [message appears]. b/kick <user> [that message will not appear as it normally did in the first place].)
-b/info servers were updated a bit.
If you do b/info server (by total accident, it will say, "Did you mean b/info servers?" without having an error come up.
This was done so the bot can remind people like its no big deal, just re-type it, and your good.
-b/poll was updated a bit.
Now when doing command b/poll without any message after that in regular, it will show a more helpful command, that is much easier to read, and gives more detail on how it works than it was before.
This was done, so people know how it works a bit more, and what the functions on the code I have written, so it makes more sense to them.

*Version 1.01.1 Release*
-Updated b/invite. We have released a new more controllable invite link, so we switched the permissions ID for it to work properly. 

*Version 1.01.2 Release*
-Fixed the b/report command because of a huge error we made. If you did not mention @ user at all, the bot would shut down. We fixed that by hiding your name when reporting a player. The staff team will still see the report, as long as the user is mentioned.

And Finally
*Version 1.02.0 Release*
-New Command! b/changelog. This will show the latest bot patch notes.
This was done so people can check the latest patch notes, right on their server. If they want to check old patch notes, you will have to join the SuperBot! Discord Server to see archived patch notes
-Changed the b/help top section. Because of the new b/changelog command, we changed it so people can see the latest patch notes to the bot
 
-discord.js is now officially installed to 12.3.1. This patch came out on August 15th, 2020.
-b/info about and b/info versions have changed because of the new discord.js update.

(Super Special Patch Notes)
-We added a new easter egg. Have fun trying to find it...

More updates coming soon...