Version 1.13.0
Huge File Change! (Read Carefully...)

--Lines 53-74--
-b/say has changed. I have been finding errors with that command, so we fixed it by adding a small intro when using b/say and nothing else after it.
(This prevents the bot from having errors)

--Lines 173-206--
-b/report has changed. I have been experimenting with some stuff. I did make a change with the SuperBot! Beta where it sections out text. So on line 189, you can change the "1" to "1, 2" in the code, and that is what I am running with that on the SuperBot! Beta for version 1.2.0 which as of right now, is currently under beta.

--Ect...--
-On some lines, you will see this.
""if(!message.content.startsWith(PREFIX) || message.author.bot)return;""
and others, you will see this
""if(!message.content.startsWith(PREFIX) || message.author.bot || message.channel.type == "dm")return;""

(message.channel.type == "dm") is on some commands because it disabled the ability for a person to use it in the direct message box with the bot. This can cause problems, and the bot can shut down, so we added it on some commands so people using the bot in direct messages, will not be able to access those commands.

-More updates coming soon...