basiccommands.js Version 1.2.0
Only 1 change occurred in this 1.2.0 update

-b/poll has changed. Now it has a new message if you want to leave a review on the bot. Otherwise, everything else is exactly the same.
(Lines 165-167).

On every update, we added a command that will tell what is new in each update. On lines 261-277, we added a command called, "changelog." When there is a new update, we will update this command every update. This will not be mentioned at all when we update basic commands for each update.